Isolate hosts list retrieval into its own module
