+++ js/searchbar/hostsSuggestionsPlugin.js
@@ -1,8 +1,7 @@
 var searchbarPlugins = require('searchbar/searchbarPlugins.js')
 var searchbarUtils = require('searchbar/searchbarUtils.js')
 var urlParser = require('util/urlParser.js')
-
-// hosts are parsed in util/urlParser
+const hosts = require('util/hosts.js')
 
 function showHostsSuggestions (text, input, event, container) {
   empty(container)
